# Conflicts:
#	.github/workflows/gradle.yml
#	build.gradle
#	src/gametest/java/net/wurstclient/gametest/WurstTest.java
#	src/main/java/net/wurstclient/commands/ModifyCmd.java
#	src/main/java/net/wurstclient/commands/PotionCmd.java

+++ src/main/java/net/wurstclient/other_features/ZoomOtf.java
@@ -53,7 +53,7 @@ public final class ZoomOtf extends OtherFeature implements MouseScrollListener
 		EVENTS.add(MouseScrollListener.class, this);
 	}
 	
-	public float changeFovBasedOnZoom(float fov)
+	public double changeFovBasedOnZoom(double fov)
 	{
 		OptionInstance<Double> mouseSensitivitySetting =
 			MC.options.sensitivity();
@@ -83,7 +83,7 @@ public final class ZoomOtf extends OtherFeature implements MouseScrollListener
 		mouseSensitivitySetting
 			.set(defaultMouseSensitivity * (1.0 / currentLevel));
 		
-		return (float)(fov / currentLevel);
+		return fov / currentLevel;
 	}
 	
 	@Override
